\name{load.mnist}
\alias{load.mnist}
\title{Load MNIST DataSet}
\usage{
load.mnist(dir)
}
\arguments{
  \item{dir}{dir of minst dataset}
}
\value{
mnist dataset train$n number of train samples train$x pix
of every train sample image train$y label of every train
sample image train$yy one-of-c vector of label of train
sample image test$n number of test samples test$x pix of
every test sample image test$y label of every test sample
image test$yy one-of-c vector of label of test sample image
}
\description{
Load MNIST DataSet
}
\author{
Xiao Rong
}
